+++ sormas-ui/src/main/java/de/symeda/sormas/ui/dashboard/statistics/DashboardStatisticsGraphicalGrowthElement.java
@@ -75,7 +75,7 @@ public class DashboardStatisticsGraphicalGrowthElement extends VerticalLayout {
 		addComponent(svgBarElement);
 	}
 
-	public void update(long count, int percentage, int growthPercentage, boolean increaseIsPositive, boolean showPercentage) {
+	public void update(int count, int percentage, int growthPercentage, boolean increaseIsPositive, boolean showPercentage) {
 		countLabel.setValue(count + (showPercentage ? " (" + percentage + " %)" : ""));
 		svgBarElement.updateSvg(percentage);
 		percentageLabel.setValue(growthPercentage != Integer.MIN_VALUE ? (growthPercentage + " %") : "");
